fixes max memory available detection on linux
